Push the night's progress
